+++ examples/call-service-panel-example/src/CallServicePanel.tsx
@@ -1,4 +1,4 @@
-import { PanelExtensionContext, RenderState } from "@foxglove/extension";
+import { PanelExtensionContext, RenderState } from "@lichtblick/suite";
 import { useCallback, useEffect, useLayoutEffect, useState, ReactElement } from "react";
 import { createRoot } from "react-dom/client";
 import ReactJson from "react-json-view";