+++ examples/custom-image-extension/src/index.ts
@@ -1,10 +1,10 @@
-import { ExtensionContext } from "@foxglove/extension";
+import { ExtensionContext } from "@lichtblick/suite";
 
 import { initExamplePanel } from "./ExamplePanel";
 
 export function activate(extensionContext: ExtensionContext): void {
   extensionContext.registerPanel({
-    name: "Foxglove Example Image Extension Panel",
+    name: "Lichtblick Example Image Extension Panel",
     initPanel: initExamplePanel,
   });
 }