+++ examples/panel-settings/src/index.ts
@@ -1,10 +1,10 @@
-import { ExtensionContext } from "@foxglove/extension";
+import { ExtensionContext } from "@lichtblick/suite";
 
 import { initExamplePanel } from "./ExamplePanel";
 
 export function activate(extensionContext: ExtensionContext): void {
   extensionContext.registerPanel({
-    name: "Foxglove Panel Settings Example",
+    name: "Lichtblick Panel Settings Example",
     initPanel: initExamplePanel,
   });
 }