+++ src/bin/create-lichtblick-extension.ts
@@ -4,9 +4,7 @@ import { program } from "commander";
 import { createCommand } from "..";
 
 program
-  .description(
-    "Creates a new Foxglove extension. Docs: https://docs.foxglove.dev/docs/visualization/extensions/introduction",
-  )
+  .description("Creates a new Lichtblick extension.")
   .showHelpAfterError()
   .argument("<name>", "Name for the new extension")
   .allowExcessArguments(false)