+++ src/bin/lichtblick-extension.ts
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 import { program, Option } from "commander";
 
-import { buildCommand, installCommand, packageCommand, publishCommand, PublishOptions } from "..";
+import { buildCommand, installCommand, packageCommand } from "..";
 import { fatal } from "../log";
 
 function main(task: Promise<void>): void {
@@ -37,20 +37,6 @@ program
   .action(({ cwd }) => {
     // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
     main(installCommand({ cwd }));
-  });
-
-program
-  .command("publish")
-  .description(
-    "Create an extensions.json entry for a released extension. This can be added to the https://github.com/foxglove/extension-registry repository",
-  )
-  .option("--foxe <foxe>", "URL of the published .foxe file")
-  .option("--cwd [cwd]", "Directory containing the extension package.json file")
-  .option("--version [version]", "Version of the published .foxe file")
-  .option("--readme [readme]", "URL of the extension README.md file")
-  .option("--changelog [changelog]", "URL of the extension CHANGELOG.md file")
-  .action((options: PublishOptions) => {
-    main(publishCommand(options));
   });
 
 program.on("command:*", ([_cmd]: string) => {