+++ template/src/ExamplePanel.tsx
@@ -1,4 +1,4 @@
-import { Immutable, MessageEvent, PanelExtensionContext, Topic } from "@foxglove/extension";
+import { Immutable, MessageEvent, PanelExtensionContext, Topic } from "@lichtblick/suite";
 import { ReactElement, useEffect, useLayoutEffect, useState } from "react";
 import { createRoot } from "react-dom/client";
 
@@ -59,7 +59,7 @@ function ExamplePanel({ context }: { context: PanelExtensionContext }): ReactEle
       <p>
         Check the{" "}
         <a href="https://foxglove.dev/docs/studio/extensions/getting-started">documentation</a> for
-        more details on building extension panels for Foxglove Studio.
+        more details on building extension panels for Lichtblick Suite.
       </p>
       <div style={{ display: "grid", gridTemplateColumns: "1fr 1fr", rowGap: "0.2rem" }}>
         <b style={{ borderBottom: "1px solid" }}>Topic</b>